complete the functionality for updating new comment instantly in the dom
