+++ src/modules/comment.js
@@ -33,6 +33,27 @@ export const createCommentBubble = ({ creation_date, comment, username }) => {
   return commentDiv;
 };
 
+export const generateComments = async (id) =>{
+  const commentWrapper = document.createElement('div');
+  commentWrapper.className = 'comment-wrapper';
+  commentWrapper.innerHTML = '';
+
+  const comments = await fetchComments(id);
+  if (comments.length <= 0) {
+    const noCommentsSpan = document.createElement('span');
+    noCommentsSpan.className = 'no-comment-span';
+    noCommentsSpan.textContent = 'Be the first to comment ... ';
+    commentWrapper.append(noCommentsSpan);
+  } else {
+    comments.forEach((commment) => {
+      commentWrapper.append(createCommentBubble(commment));
+    });
+  }
+
+  return commentWrapper
+
+}
+
 
 export const addNewComment = async (id, user, comment) => {
 
@@ -49,9 +70,8 @@ export const addNewComment = async (id, user, comment) => {
             },
             body: JSON.stringify(newComment),
       })
-      console.log(res)
       const data = await res.json()
-      console.log(data)
+      return data
   } catch (error) {
       return error
   }
