+++ src/modules/popup-details.js
@@ -1,4 +1,4 @@
-import { fetchComments, createCommentBubble } from './comment.js';
+import { fetchComments, createCommentBubble, generateComments, addNewComment } from './comment.js';
 
 export const fetchFoodDetails = async (id) => {
   const baseUrl = 'https://www.themealdb.com/api/json/v1/1/lookup.php?i=';
@@ -17,15 +17,26 @@ export const fetchFoodDetails = async (id) => {
   }
 };
 
-export const addCommentEvent = () => {
+export const addCommentEvent = async () => {
   const commentForm = document.getElementById('new-comment');
   const addCommentBtn = document.querySelector('.add-comment-btn');
-  commentForm.addEventListener('submit', (e) => {
+  commentForm.addEventListener('submit', async (e) => {
     e.preventDefault();
     const user = commentForm.elements.name
     const comment = commentForm.elements.comment
 
-    console.log(user.value, comment.value, addCommentBtn.id)
+
+    await addNewComment(addCommentBtn.id, user.value, comment.value )
+
+    const cmtSection = document.querySelector('.comments-section');
+
+  
+    let commentWrapper = await generateComments(addCommentBtn.id)
+    
+    cmtSection.innerHTML = ``
+    cmtSection.append(commentWrapper);
+
+    commentForm.reset()
   });
 };
 
@@ -139,54 +150,21 @@ export const creatPopUp = ({
   return PopUpContent;
 };
 
-const generateComments = async () =>{
-  const commentWrapper = document.createElement('div');
-  commentWrapper.className = 'comment-wrapper';
-  commentWrapper.innerHTML = '';
 
-  const comments = await fetchComments('4454');
-  if (comments.length <= 0) {
-    const noCommentsSpan = document.createElement('span');
-    noCommentsSpan.className = 'no-comment-span';
-    noCommentsSpan.textContent = 'Be the first to comment ... ';
-    commentWrapper.append(noCommentsSpan);
-  } else {
-    comments.forEach((commment) => {
-      commentWrapper.append(createCommentBubble(commment));
-    });
-  }
-
-  return commentWrapper
-
-}
 
 export const displayPopUp = async (id) => {
   const PopUpDiv = document.querySelector('.pop-up');
-  // const commentWrapper = document.createElement('div');
-  // commentWrapper.className = 'comment-wrapper';
   PopUpDiv.innerHTML = '';
-  // commentWrapper.innerHTML = '';
 
   const foodDetails = await fetchFoodDetails(id);
   PopUpDiv.append(creatPopUp(foodDetails[0]));
 
   const cmtSection = document.querySelector('.comments-section');
 
-  // const comments = await fetchComments('4454');
-  // if (comments.length <= 0) {
-  //   const noCommentsSpan = document.createElement('span');
-  //   noCommentsSpan.className = 'no-comment-span';
-  //   noCommentsSpan.textContent = 'Be the first to comment ... ';
-  //   commentWrapper.append(noCommentsSpan);
-  // } else {
-  //   comments.forEach((commment) => {
-  //     commentWrapper.append(createCommentBubble(commment));
-  //   });
-  // }
-
-  let commentWrapper = await generateComments()
+  let commentWrapper = await generateComments(id)
 
   cmtSection.append(commentWrapper);
+  console.clear()
 
   PopUpDiv.style.display = 'block';
   addCommentEvent()
